third_party/WebKit/LayoutTests/http/tests/serviceworker/fetch-script-onerror-expected.txt

CONSOLE ERROR: line 1: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: line 1: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: line 1: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: line 1: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: line 1: Uncaught ReferenceError: UndefinedReference is not defined
CONSOLE ERROR: line 1: Uncaught ReferenceError: UndefinedReference is not defined
This is a testharness.js-based test.
PASS Check script error sanitization
Harness: the test ran to completion.
